core/translate: move DELETE/UPDATE statement binding into the bind module
bind_delete_stmt and bind_update_stmt lived in delete.rs/update.rs and
called back into BindContext, so statement-level binding bounced between
modules and each wrapped its output in a DeleteBinding/UpdateBinding
struct that just re-bundled BoundDelete/BoundUpdate with the target
table metadata. Move the orchestration (database resolution, target
validation, binding) into bind.rs and fold database_id/table/or_conflict
into BoundDelete/BoundUpdate directly.

begin_write_on_database is a program-building side effect, not binding,
so it moves out of the bind phase into translate_delete and
prepare_update_plan.
