+++ core/translate/bind.rs
@@ -566,6 +566,12 @@ pub struct BoundUpdate {
     pub subquery_bindings: HashMap<ast::TableInternalId, BoundSubquery>,
     pub derived_bindings: HashMap<ast::TableInternalId, BoundSubquery>,
     pub cte_definitions: Vec<(String, CteEntry)>,
+    /// Database the target table lives in (0 = main).
+    pub database_id: usize,
+    /// The validated target table.
+    pub table: Arc<Table>,
+    /// `OR <conflict>` clause taken off the statement during binding.
+    pub or_conflict: Option<ast::ResolveType>,
 }
 
 impl BoundUpdate {
@@ -608,6 +614,10 @@ pub struct BoundDelete {
     pub tracking: BindTracking,
     pub subquery_bindings: HashMap<ast::TableInternalId, BoundSubquery>,
     pub cte_definitions: Vec<(String, CteEntry)>,
+    /// Database the target table lives in (0 = main).
+    pub database_id: usize,
+    /// The validated target table.
+    pub table: Arc<Table>,
 }
 
 impl BoundDelete {
@@ -623,6 +633,166 @@ impl BoundDelete {
             Vec::new(),
         )
     }
+}
+
+// ── Statement-level binding ──────────────────────────────────────────────
+
+/// Bind a DELETE statement up front: validate the target table and resolve
+/// all names in WHERE and RETURNING. Planning consumes the result without
+/// re-resolving anything.
+#[allow(clippy::too_many_arguments)]
+pub fn bind_delete_stmt(
+    tbl_name: &ast::QualifiedName,
+    indexed: Option<ast::Indexed>,
+    where_clause: &mut Option<Box<ast::Expr>>,
+    returning: &mut Vec<ast::ResultColumn>,
+    with: &mut Option<ast::With>,
+    resolver: &Resolver,
+    program: &mut ProgramBuilder,
+    connection: &Arc<crate::Connection>,
+) -> Result<BoundDelete> {
+    let database_id = resolver.resolve_existing_table_database_id_qualified(tbl_name)?;
+    let normalized_table_name = normalize_ident(tbl_name.name.as_str());
+    let table = validate_delete(
+        resolver,
+        &normalized_table_name,
+        database_id,
+        program,
+        connection,
+    )?;
+    let mut binder = BindContext::new(resolver, program);
+    binder.bind_delete(
+        tbl_name,
+        indexed,
+        where_clause,
+        returning,
+        with,
+        database_id,
+        table,
+    )
+}
+
+/// Bind an UPDATE statement up front: validate the target table and resolve
+/// all names in FROM/SET/WHERE/RETURNING. Planning consumes the result
+/// without re-resolving anything.
+pub fn bind_update_stmt(
+    body: &mut ast::Update,
+    resolver: &Resolver,
+    program: &mut ProgramBuilder,
+    connection: &Arc<crate::Connection>,
+    is_internal_schema_change: bool,
+) -> Result<BoundUpdate> {
+    let database_id = resolver.resolve_existing_table_database_id_qualified(&body.tbl_name)?;
+    let target_name = &body.tbl_name.name;
+    let table = match resolver.with_schema(database_id, |s| s.get_table(target_name.as_str())) {
+        Some(table) => table,
+        None => crate::bail_parse_error!("Parse error: no such table: {}", target_name),
+    };
+    if program.trigger.is_some() && table.virtual_table().is_some() {
+        crate::bail_parse_error!(
+            "unsafe use of virtual table \"{}\"",
+            body.tbl_name.name.as_str()
+        );
+    }
+    if table.btree().is_some_and(|bt| !bt.has_rowid) {
+        crate::bail_parse_error!("UPDATE of WITHOUT ROWID tables is not supported");
+    }
+    validate_update(
+        resolver.schema(),
+        body,
+        target_name.as_str(),
+        is_internal_schema_change,
+        connection,
+    )?;
+    let mut binder = BindContext::new(resolver, program);
+    binder.bind_update(body, database_id, table)
+}
+
+/// Validate the DELETE target, returning the underlying table if validation
+/// passes.
+fn validate_delete(
+    resolver: &Resolver,
+    tbl_name: &str,
+    database_id: usize,
+    program: &ProgramBuilder,
+    connection: &Arc<crate::Connection>,
+) -> Result<Arc<Table>> {
+    // Check if this is a system table that should be protected from direct writes
+    if !connection.is_nested_stmt()
+        && !connection.is_mvcc_bootstrap_connection()
+        && !crate::schema::allow_user_dml(tbl_name)
+    {
+        crate::bail_parse_error!("table {tbl_name} may not be modified");
+    }
+    let table = match resolver.with_schema(database_id, |s| s.get_table(tbl_name)) {
+        Some(table) => table,
+        None => crate::bail_parse_error!("no such table: {}", tbl_name),
+    };
+    if program.trigger.is_some() && table.virtual_table().is_some() {
+        crate::bail_parse_error!("unsafe use of virtual table \"{}\"", tbl_name);
+    }
+    if table.btree().is_some_and(|bt| !bt.has_rowid) {
+        crate::bail_parse_error!("DELETE from WITHOUT ROWID tables is not supported");
+    }
+
+    // Check if this is a materialized view
+    if resolver.schema().is_materialized_view(tbl_name) {
+        crate::bail_parse_error!("cannot modify materialized view {}", tbl_name);
+    }
+
+    // Check if this table has any incompatible dependent views
+    resolver.schema().with_incompatible_dependent_views(tbl_name, |views| {
+    if !views.is_empty() {
+        use crate::incremental::compiler::DBSP_CIRCUIT_VERSION;
+        crate::bail_parse_error!(
+            "Cannot DELETE from table '{tbl_name}' because it has incompatible dependent materialized view(s): {}. \n\
+             These views were created with a different DBSP version than the current version ({DBSP_CIRCUIT_VERSION}). \n\
+             Please DROP and recreate the view(s) before modifying this table.",
+            views.iter().fold(String::new(), |_, s| s.to_string() + ", "),
+        );
+    }
+    Ok(())
+    })?;
+    Ok(table)
+}
+
+/// Validate the UPDATE target and statement shape.
+fn validate_update(
+    schema: &crate::schema::Schema,
+    body: &ast::Update,
+    table_name: &str,
+    is_internal_schema_change: bool,
+    conn: &Arc<crate::Connection>,
+) -> Result<()> {
+    // Check if this is a system table that should be protected from direct writes
+    if !is_internal_schema_change
+        && !conn.is_nested_stmt()
+        && !conn.is_mvcc_bootstrap_connection()
+        && !crate::schema::allow_user_dml(table_name)
+    {
+        crate::bail_parse_error!("table {} may not be modified", table_name);
+    }
+    if !body.order_by.is_empty() {
+        crate::bail_parse_error!("ORDER BY is not supported in UPDATE");
+    }
+    // Check if this is a materialized view
+    if schema.is_materialized_view(table_name) {
+        crate::bail_parse_error!("cannot modify materialized view {}", table_name);
+    }
+
+    // Check if this table has any incompatible dependent views
+    schema.with_incompatible_dependent_views(table_name, |views| {
+    if !views.is_empty() {
+        use crate::incremental::compiler::DBSP_CIRCUIT_VERSION;
+        crate::bail_parse_error!(
+            "Cannot UPDATE table '{table_name}' because it has incompatible dependent materialized view(s): {}. \n\
+             These views were created with a different DBSP version than the current version ({DBSP_CIRCUIT_VERSION}). \n\
+             Please DROP and recreate the view(s) before modifying this table.",
+            views.iter().map(|view| view.as_str()).collect::<Vec<_>>().join(", "),
+        );
+    }
+    Ok(())
+    })
 }
 
 #[derive(Clone)]
@@ -3440,7 +3610,9 @@ impl<'a, G: IdGenerator> BindContext<'a, G> {
         &mut self,
         update: &mut ast::Update,
         database_id: usize,
+        table: Arc<Table>,
     ) -> Result<BoundUpdate> {
+        let or_conflict = update.or_conflict.take();
         self.with_query(|ctx| {
             // 1. Bind CTEs from WITH clause
             if let Some(with) = &mut update.with {
@@ -3530,6 +3702,9 @@ impl<'a, G: IdGenerator> BindContext<'a, G> {
                 subquery_bindings: std::mem::take(&mut ctx.subquery_bindings),
                 derived_bindings: std::mem::take(&mut ctx.derived_bindings),
                 cte_definitions,
+                database_id,
+                table,
+                or_conflict,
             })
         })
     }
@@ -3620,6 +3795,7 @@ impl<'a, G: IdGenerator> BindContext<'a, G> {
         returning: &mut Vec<ast::ResultColumn>,
         with: &mut Option<ast::With>,
         database_id: usize,
+        table: Arc<Table>,
     ) -> Result<BoundDelete> {
         self.with_query(|ctx| {
             // 1. Bind CTEs from WITH clause
@@ -3664,6 +3840,8 @@ impl<'a, G: IdGenerator> BindContext<'a, G> {
                 tracking: std::mem::take(&mut ctx.tracking),
                 subquery_bindings: std::mem::take(&mut ctx.subquery_bindings),
                 cte_definitions,
+                database_id,
+                table,
             })
         })
     }
