+++ core/translate/delete.rs
@@ -12,65 +12,17 @@ use crate::translate::subquery::{
     plan_subqueries_from_where_clause,
 };
 use crate::translate::trigger_exec::has_triggers_including_temp;
-use crate::util::normalize_ident;
 use crate::vdbe::builder::{ProgramBuilder, ProgramBuilderOpts};
 use crate::Result;
 use smallvec::SmallVec;
-use turso_parser::ast::{Expr, Limit, QualifiedName, RefAct, ResultColumn, TriggerEvent, With};
+use turso_parser::ast::{Expr, Limit, RefAct, ResultColumn, TriggerEvent};
 
 use super::plan::WhereTerm;
-
-// validate the delete statment, returning the underlying table if validation passes
-fn validate_delete(
-    resolver: &Resolver,
-    tbl_name: &str,
-    database_id: usize,
-    program: &mut ProgramBuilder,
-    connection: &Arc<crate::Connection>,
-) -> Result<Arc<Table>> {
-    // Check if this is a system table that should be protected from direct writes
-    if !connection.is_nested_stmt()
-        && !connection.is_mvcc_bootstrap_connection()
-        && !crate::schema::allow_user_dml(tbl_name)
-    {
-        crate::bail_parse_error!("table {tbl_name} may not be modified");
-    }
-    let table = match resolver.with_schema(database_id, |s| s.get_table(tbl_name)) {
-        Some(table) => table,
-        None => crate::bail_parse_error!("no such table: {}", tbl_name),
-    };
-    if program.trigger.is_some() && table.virtual_table().is_some() {
-        crate::bail_parse_error!("unsafe use of virtual table \"{}\"", tbl_name);
-    }
-    if table.btree().is_some_and(|bt| !bt.has_rowid) {
-        crate::bail_parse_error!("DELETE from WITHOUT ROWID tables is not supported");
-    }
-
-    // Check if this is a materialized view
-    if resolver.schema().is_materialized_view(tbl_name) {
-        crate::bail_parse_error!("cannot modify materialized view {}", tbl_name);
-    }
-
-    // Check if this table has any incompatible dependent views
-    resolver.schema().with_incompatible_dependent_views(tbl_name, |views| {
-    if !views.is_empty() {
-        use crate::incremental::compiler::DBSP_CIRCUIT_VERSION;
-        crate::bail_parse_error!(
-            "Cannot DELETE from table '{tbl_name}' because it has incompatible dependent materialized view(s): {}. \n\
-             These views were created with a different DBSP version than the current version ({DBSP_CIRCUIT_VERSION}). \n\
-             Please DROP and recreate the view(s) before modifying this table.",
-            views.iter().fold(String::new(), |_, s| s.to_string() + ", "),
-        );
-    }
-    Ok(())
-    })?;
-    Ok(table)
-}
 
 #[allow(clippy::too_many_arguments)]
 #[turso_macros::trace_stack]
 pub fn translate_delete(
-    binding: DeleteBinding,
+    bound: super::bind::BoundDelete,
     resolver: &Resolver,
     where_clause: Option<Box<Expr>>,
     limit: Option<Limit>,
@@ -78,21 +30,17 @@ pub fn translate_delete(
     program: &mut ProgramBuilder,
     connection: &Arc<crate::Connection>,
 ) -> Result<()> {
-    let DeleteBinding {
-        bound,
-        database_id,
-        table,
-    } = binding;
+    let database_id = bound.database_id;
+    let schema_cookie = resolver.with_schema(database_id, |s| s.schema_version);
+    program.begin_write_on_database(database_id, schema_cookie)?;
 
     let (mut delete_plan, mut bound_subqueries) = prepare_delete_plan(
         program,
         resolver,
-        table,
         where_clause,
         limit,
         returning,
         connection,
-        database_id,
         bound,
     )?;
 
@@ -171,76 +119,23 @@ pub fn translate_delete(
     Ok(())
 }
 
-/// Output of the statement-level bind phase for DELETE, produced by
-/// [bind_delete_stmt] before planning starts.
-pub struct DeleteBinding {
-    pub bound: super::bind::BoundDelete,
-    pub database_id: usize,
-    pub table: Arc<Table>,
-}
-
-/// Bind a DELETE statement up front: validate the target table and resolve
-/// all names in WHERE and RETURNING. Planning consumes the result without
-/// re-resolving anything.
-#[allow(clippy::too_many_arguments)]
-pub fn bind_delete_stmt(
-    tbl_name: &QualifiedName,
-    indexed: Option<turso_parser::ast::Indexed>,
-    where_clause: &mut Option<Box<Expr>>,
-    returning: &mut Vec<ResultColumn>,
-    with: &mut Option<With>,
-    resolver: &Resolver,
-    program: &mut ProgramBuilder,
-    connection: &Arc<crate::Connection>,
-) -> Result<DeleteBinding> {
-    let database_id = resolver.resolve_existing_table_database_id_qualified(tbl_name)?;
-    let normalized_table_name = normalize_ident(tbl_name.name.as_str());
-    let table = validate_delete(
-        resolver,
-        &normalized_table_name,
-        database_id,
-        program,
-        connection,
-    )?;
-
-    let schema_cookie = resolver.with_schema(database_id, |s| s.schema_version);
-    program.begin_write_on_database(database_id, schema_cookie)?;
-
-    // Bind phase: resolve all names in WHERE and RETURNING up front.
-    let mut binder = super::bind::BindContext::new(resolver, program);
-    let bound = binder.bind_delete(
-        tbl_name,
-        indexed,
-        where_clause,
-        returning,
-        with,
-        database_id,
-    )?;
-
-    Ok(DeleteBinding {
-        bound,
-        database_id,
-        table,
-    })
-}
-
 #[allow(clippy::too_many_arguments)]
 #[turso_macros::trace_stack]
 pub fn prepare_delete_plan(
     program: &mut ProgramBuilder,
     resolver: &Resolver,
-    table: Arc<Table>,
     where_clause: Option<Box<Expr>>,
     limit: Option<Limit>,
     mut returning: Vec<ResultColumn>,
     connection: &Arc<crate::Connection>,
-    database_id: usize,
     mut bound: super::bind::BoundDelete,
 ) -> Result<(
     Plan,
     rustc_hash::FxHashMap<turso_parser::ast::TableInternalId, super::bind::BoundSubquery>,
 )> {
     let schema = resolver.schema();
+    let database_id = bound.database_id;
+    let table = bound.table.clone();
 
     let btree_table_for_triggers = table.btree();
     let indexes = schema.get_indices(table.get_name()).cloned().collect();
